Refactor `guid demote its use.
`guid` was originally envisioned as a semantic ID (hash of an entry) to
make mapping of entries easier for bulk importers.
This is unnecessary and superfluous as external importers can maintain
such maps externally without having to bring over that layer into the
DB.

This refactor changes `guid` from a semantic hash to a random UUID
and replaces its use in queries with numeric primary keys. The `guid`
field will only be used in public APIs such as new entry and correction
submissions.