Replace `chi` router with `labstack/echo` HTTP server.
This is a big refactor. With admin templates, static files, HTML site
templates and other mixed media types, `echo` is a better out of the box
solution for handling routes and rendering responses. It eliminates the
boilerplate chi required as the complexity has grown.

(Optional) HTML site templates are bound to echo's built in renderer
and admin templates are handled separately in `adminPage()`.