Make entry `content` an array `content[]`. This is a breaking change.
This patch refactors the `entries.content` table field from TEXT to TEXT[] to
support multiple content values per entry (e.g., alternative spellings,
transliterations, or script variations of the same word).

This was added after working with the Kurup Malayalam thesaurus, where each
"definition" entry is actually a series of synonyms. Rather than joining
everything with a comma and dumping in `content`, it's better to provision for
multiple values.

- Break to version v4.0.0
- Database schema: `content TEXT` to `content TEXT[]`
- APIs: `content` is now an array in all JSON responses and reqest payloads